requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
aiohttp>=3.9.1
PyPDF2>=3.0.0
openrouter>=0.3.0 